update session types and uri utils
