+++ packages/utils/test/uri.spec.ts
@@ -1,25 +1,19 @@
 import "mocha";
 import { expect } from "chai";
 import { UriParameters } from "@walletconnect/types";
-import { safeJsonStringify } from "@walletconnect/safe-json";
 
 import { formatUri, parseUri } from "../src";
-import { TEST_KEY_PAIRS, TEST_PAIRING_TOPIC, TEST_RELAY_OPTIONS } from "./shared";
+import { TEST_SYMMETRIC_KEY, TEST_PAIRING_TOPIC, TEST_RELAY_OPTIONS } from "./shared";
 
 const TEST_URI_PARAMS: UriParameters = {
   protocol: "wc",
   version: 2,
   topic: TEST_PAIRING_TOPIC,
-  publicKey: TEST_KEY_PAIRS["A"].publicKey,
-  controller: false,
+  symKey: TEST_SYMMETRIC_KEY,
   relay: TEST_RELAY_OPTIONS,
 };
 
-const TEST_URI_STRING = `${TEST_URI_PARAMS.protocol}:${TEST_URI_PARAMS.topic}@${
-  TEST_URI_PARAMS.version
-}?controller=${TEST_URI_PARAMS.controller}&publicKey=${
-  TEST_URI_PARAMS.publicKey
-}&relay=${encodeURIComponent(safeJsonStringify(TEST_URI_PARAMS.relay))}`;
+const TEST_URI_STRING = `${TEST_URI_PARAMS.protocol}:${TEST_URI_PARAMS.topic}@${TEST_URI_PARAMS.version}?relay#protocol=${TEST_RELAY_OPTIONS.protocol}&symKey=${TEST_SYMMETRIC_KEY}`;
 
 describe("URI", () => {
   it("formatUri", () => {
